Project report with insights and screenshot of the actual graphs
The aim related to analyzing “Consumer complaint related to Financial Products and Services” is to find out pattern which best describes some of the common complaints that consumers are facing and make sure that responsible financial services are taking a corrective action to avoid them in near future. By publishing the insights of this project, people of United States will be more aware about the possible risk involved in taking a particular financial services from a Bank or lender. There are more than 700,000 entries and this number is keep on increasing as CFPB is adding more number of complaints into the database every day. These insights are quantifiable measures to determine which financial services is receiving less number of complaints which will help general public to make a firm decision before choosing a financial service/provider for taking a mortgage or student loan. The article on home owners owe the most gives us an idea  (Jones) regarding which states are under more mortgage in United States but it lacks consumer response and feedback received from the financial services.